rerunning cells with changes in opt pkg using classes, refactoring... in 4.5
